Remove constants "unit" tests.
Unit tests are meant to test logic, not declarations. This is why we
don't test our HTML and CSS code. Unit tests express contractual, behavioral
obligations of modules. Constants do not correspond to contracts.

This test needs to go, so I'm removing it.

It seems to me that here we're simply trying to increase the coverage
metric and this is hurting the quality of our tests. Let's keep in mind
that coverage is a measure that helps us improve, but the end goal is
better testing.

Testing that constants are equal to what they are is unnecessary
duplication and doesn't add any value to the tests. Furthermore, it
violates DRY, as every constant update must come with a test update.
This is true for all testing, but in this case the same thing is
expressed in the test and in the implementation and is not a form of
contractual obligation.

If you must, you can test that the constants exist and are of specified
types. Feel free to readd that in case you consider it a necessity.