Updates to documentation and .deb building
Updates to build .deb

Now builds a .deb and .rpm capable of using the new openbazaar launcher.

Had to modify the bash launcher so that it changes its current directory
over to openbazaar's installation folder in linux. May have to add
checks for OSX later, although I believe in OSX, this script will be
irrelevant as the application will be launched via .app launcher.